# fuse: Better inode generation