--- a/internal/fuse/fuse_test.go
+++ b/internal/fuse/fuse_test.go
@@ -118,7 +118,7 @@ func TestFuseFile(t *testing.T) {
 	}
 	root := &Root{repo: repo, blobCache: bloblru.New(blobCacheSize)}
 
-	inode := fs.GenerateDynamicInode(1, "foo")
+	inode := inodeFromNode(1, node)
 	f, err := newFile(root, inode, node)
 	rtest.OK(t, err)
 	of, err := f.Open(context.TODO(), nil, nil)
@@ -161,8 +161,8 @@ func TestFuseDir(t *testing.T) {
 		ChangeTime: time.Unix(1606773732, 0),
 		ModTime:    time.Unix(1606773733, 0),
 	}
-	parentInode := fs.GenerateDynamicInode(0, "parent")
-	inode := fs.GenerateDynamicInode(1, "foo")
+	parentInode := inodeFromName(0, "parent")
+	inode := inodeFromName(1, "foo")
 	d, err := newDir(root, inode, parentInode, node)
 	rtest.OK(t, err)
 
@@ -218,4 +218,41 @@ func testTopUIDGID(t *testing.T, cfg Config, repo restic.Repository, uid, gid ui
 	rtest.OK(t, err)
 	rtest.Equals(t, uint32(0), attr.Uid)
 	rtest.Equals(t, uint32(0), attr.Gid)
+}
+
+func TestInodeFromNode(t *testing.T) {
+	node := &restic.Node{Name: "foo.txt", Type: "chardev", Links: 2}
+	ino1 := inodeFromNode(1, node)
+	ino2 := inodeFromNode(2, node)
+	rtest.Assert(t, ino1 == ino2, "inodes %d, %d of hard links differ", ino1, ino2)
+
+	node.Links = 1
+	ino1 = inodeFromNode(1, node)
+	ino2 = inodeFromNode(2, node)
+	rtest.Assert(t, ino1 != ino2, "same inode %d but different parent", ino1)
+}
+
+var sink uint64
+
+func BenchmarkInode(b *testing.B) {
+	for _, sub := range []struct {
+		name string
+		node restic.Node
+	}{
+		{
+			name: "no_hard_links",
+			node: restic.Node{Name: "a somewhat long-ish filename.svg.bz2", Type: "fifo"},
+		},
+		{
+			name: "hard_link",
+			node: restic.Node{Name: "some other filename", Type: "file", Links: 2},
+		},
+	} {
+		b.Run(sub.name, func(b *testing.B) {
+			b.ReportAllocs()
+			for i := 0; i < b.N; i++ {
+				sink = inodeFromNode(1, &sub.node)
+			}
+		})
+	}
 }