@greatroar 2022-11-11
There's TestArchiverAbortEarlyOnError again.