@MichaelEischer 2022-11-12
I have to admit I'm somewhat confused by the inode handling. My assumption was that different files should result in different inodes. However, judging from `fs.GenerateDynamicInode` that's not necessarily the case. What worries me is that e.g. 100 snapshots with 40million files each (given the repository sizes that have been mentioned every now and then that's a realistic repository size) result in a 50% chance of an inode collision. Do you know which identifier is internally used in the fuse library to track files/directories?