@greatroar 2022-11-13
Not the inode number, AFAICT. I just rigged inodeFrom(Name|Node) to always return 2, and the mount point still worked fine, with sha256sum distinguishing between the files properly.