@greatroar 2022-11-13
I just realized that with this change, hard-linked files may also share their inode across snapshots. Personally, I find that acceptable, but if it isn't, we need to hash in the snapshot id as well. XOR with the first 64 bits of that id [should be enough](https://security.stackexchange.com/a/34797).