@greatroar 2022-11-16
Looks like fs.GenerateDynamicInode doesn't allocate for one of these filenames because it's less than [tmpstringbufsize = 32](https://github.com/golang/go/blob/b1678e508bf04b32fcd8153d09c39ff25b51d287/src/cmd/compile/internal/walk/walk.go#L21), in which case the byte slice for it is allocated on the stack.